Lab3_AI_Agent/dashboard/requirements.txt
streamlit>=1.35.0
google-cloud-firestore>=2.16.0
google-cloud-aiplatform>=1.50.0
pandas>=2.0.0
